sharepoint-migration-tool/sharepoint-cloud-tool/requirements.txt
openpyxl==3.1.2
Office365-REST-Python-Client==2.4.3